Separate auth methods from public API
